requirements.txt
streamlit
pandas
numpy
scikit-learn
pickle-mixin
mysql-connector-python